resume_builder/requirements.txt
langchain==1.0.5
langchain-ollama==1.0.0
langchain-chroma==1.0.0
pandas==2.3.3
PyYAML==6.0.3
dataclasses-json==0.6.7
chromadb==1.3.4
sentence-transformers==5.1.2
watchdog==6.0.0